pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "catmoro"
version = "0.1.0"
description = "Catastrophe modeling sandbox."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "typer>=0.12",
]

[project.scripts]
catmoro = "catmoro.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
